EmailService y otras modificaciones

+++ FedeteriAPI/Models/UsuarioIn.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public int DNI { get; set; }
+        public long DNI { get; set; }
         public string Email { get; set; }
-        public int Celular { get; set; }
+        public long Celular { get; set; }
         public string Contrasena { get; set; }
         public string Nacimiento { get; set; }
