+++ FedeteriAPI/Models/UsuarioOut.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public int DNI { get; set; }
+        public long DNI { get; set; }
         public string Email { get; set; }
 
         public UsuarioOut(UsuarioIn usuario)
